+++ svelte/demos/cases/SubPanels.svelte
@@ -1,6 +1,6 @@
 <script>
 	import { getData, sectionItems, onlySectionItems } from "../data";
-	import { Form, registerEditorItem } from "../../src";
+	import { Editor, registerEditorItem } from "../../src";
 
 	import { Comments } from "wx-svelte-comments";
 	import { Tasklist } from "wx-svelte-tasklist";
@@ -19,19 +19,19 @@
 <div class="variations">
 	<div class="bg-top">
 		<h3>Normal sections</h3>
-		<Form items={items1} {values} />
+		<Editor items={items1} {values} topBar="{false}}" />
 	</div>
 	<div class="bg-top">
 		<h3>Exclusive mode</h3>
-		<Form items={items2} {values} />
+		<Editor items={items2} {values} topBar={false} />
 	</div>
 	<div class="bg">
 		<h3>Normal sections</h3>
-		<Form items={items3} {values} />
+		<Editor items={items3} {values} topBar={false} />
 	</div>
 	<div class="bg">
 		<h3>Accordion mode</h3>
-		<Form items={items4} {values} />
+		<Editor items={items4} {values} topBar={false} />
 	</div>
 </div>